Move user facing field tests around
